requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.7.0
psycopg2-binary==2.8.6
pytest==6.2.2
testcontainers==3.2.0
requests==2.25.1
httpretty==1.0.5
